Extract tests for Either
